requirements.txt
absl-py==0.15.0
astunparse==1.6.3
Babel==2.1.1
backcall==0.2.0
bio==1.5.2
biopython==1.79
biothings-client==0.3.1
bleach==4.1.0
blessed==1.20.0
bokeh==2.3.3
brokenaxes==0.4.2
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
circlify==0.15.0
clang==5.0
clize==3.0
cloudpickle==2.2.1
colorcet==3.0.1
cycler==0.11.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
debtcollector==0.10.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dna-features-viewer==3.1.3
einops==0.4.1
entrypoints==0.4
et-xmlfile==1.1.0
flatbuffers==1.12
funcsigs==0.4
gast==0.4.0
general==0.0.3.dev9
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.48.2
h5py==3.1.0
holoviews==1.14.9
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
iso8601==0.1.11
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
monotonic==0.4
msgpack-python==0.4.6
mygene==3.2.2
natsort==5.0.2
nest-asyncio==1.5.6
netaddr==0.7.18
netifaces==0.10.4
numba==0.53.1
numpy==1.19.5
nvidia-ml-py==12.560.30
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
openpyxl==3.1.3
opt-einsum==3.3.0
oslo.config==2.6.0
oslo.context==0.8.0
oslo.i18n==2.7.0
oslo.log==1.13.0
oslo.serialization==1.11.0
oslo.utils==2.7.0
packaging==21.3
pandas==1.1.5
panel==0.12.1
param==1.13.0
parso==0.7.1
pbr==1.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.18.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==6.1.0
ptflops==0.7.1.2
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyct==0.4.8
pyfaidx==0.7.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyinotify==0.9.6
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2024.1
pyviz-comms==2.2.1
PyYAML==3.11
pyzmq==19.0.2
rdkit-pypi==2021.9.4
requests==2.27.1
requests-oauthlib==2.0.0
rna-fm==0.2.2
rsa==4.9
scikit-learn==0.24.2
scipy==1.1.0
seaborn==0.11.2
sh==1.11
shap==0.41.0
sigtools==1.0
simplegeneric==0.8.1
six==1.15.0
slicer==0.0.7
stevedore==1.9.0
tenacity==8.2.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094701020/work
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==3.7.4.3
umap==0.1.1
urllib3==1.26.18
ViennaRNA==2.5.0a5
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.12.1
xgboost==1.5.2
xlrd==2.0.1
zipp==3.6.0
